kernal/requirements.txt
numpy==2.2.3
openai==1.64.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
lxml_html_clean==0.4.1
pathspec==0.12.1
psutil==7.0.0
fastapi==0.115.11
pyttsx3==2.98
python-docx==1.1.2
PyPDF2==3.0.1
openpyxl==3.1.5